requirements.txt

azure-ai-projects
azure-identity
python-dotenv
databricks-sdk
pandas
asyncio
flet
